--HG--
branch : sounddata-samplecount

+++ src/modules/sound/SoundData.cpp
@@ -152,9 +152,9 @@ int SoundData::getSampleCount() const
 	return (size/channels)/(bits/8);
 }
 
-int SoundData::getDuration() const
+float SoundData::getDuration() const
 {
-	return size/(channels*sampleRate*bits/8);
+	return float(size) /(channels*sampleRate*bits/8);
 }
 
 void SoundData::setSample(int i, float sample)
